+++ openmc_dagmc_wrapper/__init__.py
@@ -2,10 +2,7 @@ from .utils import create_material, diff_between_angles
 
 from .Geometry import Geometry
 from .Materials import Materials
-from .Tally import (
-    Tally,
-    MeshTally3D,
-    MeshTallies3D,
-)
+from .Tally import Tally
+
 from .tallies import *
 from .Settings import FusionSettings
