+++ openmc_dagmc_wrapper/tallies/regular_mesh_tallies/__init__.py
@@ -1,1 +1,2 @@
 from .mesh_tally_2D import MeshTally2D, MeshTallies2D
+from .mesh_tally_3D import MeshTally3D, MeshTallies3D
